+++ benches/primitives.rs
@@ -1,10 +1,11 @@
 use std::time::Duration;
 
-use anon_sym_ake::protocol::{
-    ccpake::{ccapke_dec, ccapke_enc},
-    prf::prf,
-    supported_algs::{get_kem_algorithm, get_signature_algorithm},
-    utils::get_random_key32,
+use anon_sym_ake::{
+    common::{prf::prf, utils::get_random_key32},
+    pq::{
+        ccapake::{ccapke_dec, ccapke_enc},
+        supported_algs::{get_kem_algorithm, get_signature_algorithm},
+    },
 };
 use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion};
