--- src/protocol/mod.rs
@@ -1,10 +1,0 @@
-pub mod ccpake;
-pub mod client;
-pub mod commitment;
-pub mod config;
-pub mod pke;
-pub mod prf;
-pub mod protocol;
-pub mod server;
-pub mod supported_algs;
-pub mod utils;
